move MirrorConfig to config package
