+++ src/test/groovy/io/seqera/wave/service/mirror/strategy/DockerMirrorStrategyTest.groovy
@@ -22,7 +22,7 @@ import spock.lang.Specification
 
 import java.nio.file.Path
 
-import io.seqera.wave.service.mirror.MirrorConfig
+import io.seqera.wave.configuration.MirrorConfig
 
 /**
  *
